src/backend/queue_agent/requirements.txt
aws-xray-sdk==2.12.0
boto3==1.33.2
botocore
certifi
charset-normalizer==3.2.0
idna==3.4
jmespath==1.0.1
Pillow==10.2.0
python-dateutil==2.8.2
requests==2.31.0
retry-requests==2.0.0
s3transfer>=0.8.0
six==1.16.0
urllib3>=2.2.0
wrapt==1.15.0
aiohttp-client-cache==0.8.2
aioboto3>=12.3.0
websocket-client==1.7.0
python-magic>=0.4.27